Support keyword filtering in fsharp-reference.com
* Add search box above menu to filter menu items and keywords associated
  with them
* Turn src/ into an F# project
* Autogenerate index.js menu for fsharp-reference
* Turn config and frontmatter into records instead of maps
* Move content for fsharp-reference to align with sections in the menu

+++ clean.ps1
@@ -1,3 +1,3 @@
-rm -r output/fsharp-reference
-rm -r output/electricvisions
-rm -r output/matter-game
+if (Test-Path 'output/fsharp-reference') { rm -r output/fsharp-reference }
+if (Test-Path 'output/electricvisions') { rm -r output/electricvisions }
+if (Test-Path 'output/matter-game') { rm -r output/matter-game }
